widen the rank exponent to u64: every reachable depth is representable
Rank.exp goes u32 -> u64 through every seam: raw_parts/from_raw/
encode_parts signatures, the from_raw shift min (now a plain u64 min),
the sum fold's running exponent, and content_bits. Base gains u64 shift
forms (checked conversion to the backend's usize; the shift meter
widens its rhs so every shift records the same total through either
width).

The decoder's fraction-depth cap dissolves rather than widens: frac_len
is counted from bits actually read, so it never exceeds the input's own
bit count and always fits the u64 exponent - no rejection genre
remains for deep fractions, and the format's one representation bound
is the integral mantissa's 2^64 bits. Canonical bytes of every
reachable value are unchanged (goldens, the exhaustive lex==Ord sweep,
and the wire snapshots all pass byte-identical).

The skyline query kernels' depth guards become total conversions
(depth counts levels of streams held in memory, so it fits u64), and
the depth>u32::MAX # Panics prose is excised everywhere it propagated;
the deep-spine 3-bits-per-level pin is re-denominated as the grammar's
depth-to-size exchange rate.

+++ crates/before/src/oracle/version.rs
@@ -220,7 +220,7 @@ impl Version {
 
     /// The raw `(numerator, exponent)` area fold, in subtree-relative units
     /// (this subtree's interval has width 1).
-    fn rank_raw(&self) -> (Base, u32) {
+    fn rank_raw(&self) -> (Base, u64) {
         match self {
             Version::Leaf(n) => (n.clone(), 0),
             Version::Node(n, l, r) => {
